nilfs2: add missing blkdev_issue_flush() to nilfs_sync_fs()
Under normal circumstances nilfs_sync_fs() writes out the super block,
which causes a flush of the underlying block device.  But this depends on
the THE_NILFS_SB_DIRTY flag, which is only set if the pointer to the last
segment crosses a segment boundary.  So if only a small amount of data is
written before the call to nilfs_sync_fs(), no flush of the block device
occurs.

In the above case an additional call to blkdev_issue_flush() is needed. 
To prevent unnecessary overhead, the new flag nilfs->ns_flushed_device is
introduced, which is cleared whenever new logs are written and set
whenever the block device is flushed.  For convenience the function
nilfs_flush_device() is added, which contains the above logic.

Signed-off-by: Andreas Rohner <andreas.rohner@gmx.net>
Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>