common: dma-mapping: Store page array in vm_struct
Commit 54329ac (common: dma-mapping: introduce common remapping functions)
factored out common code for remapping arrays of pages.  The code before
the refactor relied on setting area->pages with the array of mapped pages
for easy access later.  The refactor dropped this, breaking parts of the
ARM DMA API.  Fix this by setting the page array in the same place.

Signed-off-by: Laura Abbott <lauraa@codeaurora.org>
Reported-by: Mitchel Humpherys <mitchelh@codeaurora.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>